Support building Python binaries for a specific stack
Previously it was only possible to trigger Python binary builds for
all stacks supported by the Python version specified in the GitHub
Actions workflow request.

Now, it's possible to select a single stack, which will allow for easier
rebuilding of binaries for new stacks (such as Heroku-24, when support
for it is added in a later PR), without having to rebuild (and overwrite on
S3) all existing stacks at the same time.

GUS-W-15570758.