Build Python 3.10 runtimes for Heroku-24 too
Expands the Heroku-24 Python version support to include Python 3.10,
in addition to the existing Python 3.11 and 3.12.

This will give more Python version options to CNB users that need ARM64
support, given that it will only available be on Heroku-24+.

I've not added a changelog entry since Heroku-24 is not yet released, so
no one is using it anyway, and this is covered by the existing Heroku-24
changelog entry (the previous release of the buildpack will be able to
use these binaries from S3 too).

GUS-W-15571102.