Remove unnecessary `pipenv-to-pip` script
Since its only used to export `Pipfile.lock` to `requirements.txt`, however,
when using Pipenv the requirements file is never used (in a much older iteration
of cache invalidation, the `requirements-installed.txt` copy of the file was used).

This also means the vendor directory no longer needs to be on `PATH` (it shouldn't
have been added to the path exposed to the user-facing parts of the buildpack in the
first place).