Improve `make run`
The `make run` target provides a quick way to execute the buildpack
locally in a container when developing on the buildpack.

Previously it executed only `bin/{detect,compile,report}`.

Now it also runs `bin/release`.

In addition, the arguments passed to `bin/detect` have been corrected,
so that they now match the Buildpack API (where only `BUILD_DIR` is passed
to the script):
https://devcenter.heroku.com/articles/buildpack-api#bin-detect