Switch to a new output helper for error messages
The existing logging functions in `bin/utils` don't support being
passed multi-line output.

Upcoming PRs are going to be adding a number of new error
messages, so this adds a new output module under `lib/` (that
supports multi-line output and also uses colour) and switches
the existing buildpack error messages to it.

GUS-W-16808943.