Add support for Python 3.9 on Heroku-24
Previously Python 3.9 was only supported on the Heroku-22
stack and older, since it reaches EOL in Oct 2025.

However, adding support for Python 3.9 to Heroku-24 will make
it easier for users upgrading stacks, and also give more Python
version options to CNB users that need ARM64 support, given
that multi-arch is only available for Heroku-24+. It also simplifies
the version error checks we need to perform in the upcoming
Python version handling refactor.

GUS-W-16917997.