Fix leftover output::error args usages (#1671)
In #1666 the API for `output::error` was simplified, such that it
only supports the message being passed via stdin (since that's
the pattern used when passing a multi-line HERE-doc, and we
want to encourage detailed error messages).

However, there were a couple of remaining instances not switched
over to the stdin API.

These usages are in codepaths that cannot be tested via integration
tests since they are unreachable unless there was an internal
buildpack bug. (And Shellcheck also didn't spot these instances of
unconsumed passed args.)

GUS-W-17048951.