Fix using Poetry with outdated Python patch versions
The existing Poetry bootstrap process added in #1682 used the pip wheel
bundled within the Python stdlib.

This required use of pip's `--python` option, which was added to pip in
v22.3 in 2022. All of the major Python versions we support have been
updated to that pip version or newer, the older patch releases of some
of those major Python versions can contain pip versions that are older.

Whilst we strongly recommend upgrading to newer patch releases (since
older versions are missing security updates and so likely insecure),
we still want to support using Poetry on these versions, so the I've
adjusted the bootstrap process to no longer use `--python`.