Improve NLTK downloader log output and error messages
Improves the build log output, error messages and buildpack
failure metrics for the NLTK downloader feature, and adds a
test for the app having an invalid `nltk.txt` (since previously
only the successful cases were tested).

This should be the last failure type seen in the wild that doesn't
emit a `failure_reason` metric.

GUS-W-8059892.