Include dotfiles in error message file listings
The error messages shown when buildpack detection fails
(or when detection passes, but no required package manager
files are found later) include a file listing of the project to help
with debugging, (eg to make misspellings or having forgotten
to commit files more obvious.)

Previously these file listings didn't include dotfiles, which
meant not all project files would be shown. This is particularly
relevant given that the the newly supported `.python-version`
file is a dotfile.