Update changelog links and tests after Git tag archival
The Git tags for buildpack releases v208 to v265 have been
archived (renamed from `vNNN` to `archive/vNNN`) for
the reasons in #1699 (similar to the process performed in
the past for release v207 and older).

As such, the changelog compare URLs need updating, as do
the test fixtures that test building an app whose last (cached)
build was performed using an older version of the buildpack.

Closes #1699.