Remove support for Heroku-20
Since the Heroku-20 stack has reached end-of-life, and as such builds
using it are no longer supported by the Heroku build system:
https://devcenter.heroku.com/changelog-items/3230

This fixes the integration tests failing in CI for the Heroku-20 stack,
due to the build system now (as expected) rejecting the jobs.

Any non-Heroku consumers of this buildpack that wish to continue using
the Heroku-20 stack should pin to the previous version of this
buildpack.

GUS-W-14707057.