Start rename of `python_version_reason` metric to `python_version_origin`
So that the metric name reflects the name used within the codebase.

I've left the old metric name in for now, so that existing queries continue
to work. Once all old data has aged out that doesn't have the new metric
name, I can migrate the queries over to the new name.