Fix handling of version files that contain CRLFs
From looking at metrics, I noticed that the `cat --show-nonprinting`
approach added in #1783 to aid the visualisation of invisible characters
(such as ASCII control characters) also escapes carriage return
characters (to `^M`), which we don't want.

Instead, `sed` is now used which replaces anything not matching the
`:print:` and `:space:` groups with the Unicode substitution character
(`�`) - which means the classic buildpack now also matches the CNB's
behaviour:
https://github.com/heroku/buildpacks-python/blob/f2fdd00edf0f63b298cf88c82377885c88666440/src/python_version_file.rs#L16-L19

For a mapping of what `:print:` and `:space:` cover, see the chart at
the bottom of this page:
https://en.cppreference.com/w/cpp/string/byte/isprint

GUS-W-18225347.