Port uv test improvements to other package manager tests (#1834)
Several improvements were made to uv's package manager related
tests when adding support for uv. This ports those improvements
to the tests for the other package managers too.

A couple of the new tests have to be marked as skipped, due to bugs
that they exposed upstream in Poetry and Pipenv.

GUS-W-18319828.
GUS-W-18319833.