Work around Pipenv bug when using `--system`
After moving Pipenv to its own virtual environment in #1840 (to fix
leaking Pipenv's dependencies into the app environment), apps that
explicitly depend on a package that happen to also be a Pipenv
dependency were seeing `ModuleNotFoundError` failures.

For example:

```
ModuleNotFoundError: No module named 'certifi'
```

It appears that Pipenv's `--system` feature has a bug whereby it sees
its own dependencies (installed in its own environment) as already
installed (even though they aren't actually installed in the app
environment), so incorrectly skips installing them.

In general, it seems Pipenv's support for `--system` is pretty buggy:
https://github.com/pypa/pipenv/issues?q=is%3Aissue%20state%3Aopen%20label%3A%22--system%20flag%22
(for example, it doesn't work with `pipenv run` either)

The classic Python buildpack doesn't currently use virtual environments
(since it installs a dedicated Python install for the app, that's
already separate from system Python), and switching to them would
be a significant change that may cause other types of breakage
(for example, to apps that hardcode the path to Python).

As such for now, we need to continue to use `pipenv install --system`
(as the buildpack has done for some time), but I have found the bug
can be worked around if we set `VIRTUAL_ENV` to the Python install
location, to trick Pipenv into thinking it's operating in a virtual
environment.

The Pipenv tests have been updated to confirm that an explicit
`certifi` install now works.

GUS-W-19118553.