Speed up cache restoration
By moving the Python directory instead of copying, which internally
performs a rename instead of a copy if the cache directory is on the
same filesystem mount as the build directory (which is the case for
standard builds).

This should be particularly beneficial for apps with a large number
of Python dependencies (and so a large site-packages) - where we
see cache restoration taking up to 30 seconds in Honeycomb (P99.9).

GUS-W-19144883.