Stop installing SQLite headers and CLI
For historical reasons, the classic Python buildpack has installed the
`libsqlite3-dev`, `libsqlite3-0` and `sqlite3` packages as part of the
build.

However, these are not actually required for the Python `sqlite3`
stdlib module to work (since the headers are installed in the
`Dockerfile` used specifically for those builds):
https://docs.python.org/3/library/sqlite3.html

There is also a `pysqlite3` package on PyPI that uses the headers,
however:
1. It's virtually identical to the stdlib module with Python, and so
   most apps should just use the stdlib's `sqlite3` module instead.
2. For apps that absolutely must use `pysqlite3`, the maintainer
   of that package makes a pre-compiled binary version of the
   package available that can be used without the need for headers:
   `pysqlite3-binary`
3. For apps that really want to install `pysqlite3` from source (or
   have some other need for the SQLite headers), they can be installed
   using the APT buildpack by adding `libsqlite3-dev` and `libsqlite3-0`
   to the `Aptfile`.

Metrics show that only two apps built in the last 60 days use the
`pysqlite3` package, confirming that including the headers by default
does not make sense.

This just leaves the SQLite CLI, which:
- Won't be very commonly used, given that most apps don't use SQLite
  due to the ephemeral storage on dynos
- Is outside the scope of what this buildpack should be installing,
  and is why the APT buildpack exists

...and so is also something the buildpack shouldn't be installing
itself. Instead, any apps that need the SQLite CLI should use the
APT buildpack and add `sqlite3` to their `Aptfile`.

By removing the installation of these packages, we:
- Speed up the build (metrics show the install time over last
  60 days has a P50 of 4.2s, average of 6.4s, P99.9 of 667s)
- Improve build reliability, since it no longer depends on the
  uptime of the Ubuntu APT mirror
- Improve parity with the Python CNB, which doesn't install the
  SQLite headers or CLI either
- (As an added bonus) Reduce the slug size slightly

The buildpack already doesn't install these packages when using
Python 3.13 (which is the buildpack's default Python version),
or when using Poetry or uv. As such, this change only affects
apps using Python 3.12 or older, with pip or Pipenv.

Closes #1703.
GUS-W-18915614.