Fix the stack check error message for Heroku-20
The `heroku-20` entry was removed completely in #1778
rather than moving it to the EOL branch of the conditional,
causing the "stack not recognised" message variant to be
shown, instead of the "stack no longer supported" version.

This error messages can never occur on Heroku (since our
build system can only scheduled builds for supported stacks),
but these cases can be hit for non-Heroku uses of this
buildpack (such as Dokku).

GUS-W-19274871.