Use EPOCHREALTIME in build_data timing logic
Built-in Bash (5.0+) variable, a float with microseconds precision. Faster, simpler, and more portable (no longer requires GNU date).

Logging it with six digits worth of precision instead of the previous three, but a float is a float ;)

(without a precision indicator, awk gives the usual hand-wavy float result of possibly a few more digits, we just discard that)

GUS-W-19404492