Fix collectstatic and NLTK downloader for apps with `CACHE_DIR` set (#1888)
By default, user-provided env vars (app config vars) are not loaded into
the buildpack environment by the build system, and instead buildpacks
must load them explicitly when required:
https://devcenter.heroku.com/articles/buildpack-api#bin-compile

In general, this buildpack (like many of the others) tries to limit the
places where user-provided env vars are loaded, in order to prevent them
from breaking the build process.

However, there are key parts of the build where they must be loaded, so
that the app runs as expected.

For example, when running the Django collectstatic command (to generate
the app's static assets), the Django app's config will be loaded, and
parts of it may reference/depend upon the app config vars.

For some time, the entire collectstatic step was run in a subshell with
the user config vars loaded, which meant there was a chance for env var
shadowing.

In particular, if an app had the env var `CACHE_DIR` set, it would cause
the build data store writes to fail. Until the recent `bin/report`
refactor these failing writes were silently ignored, however, those
failures now (intentionally) are surfaced, which caused the build to
fail as seen in #1886.

The Django collectstatic and NLTK downloader steps are the last two
parts of the buildpack that haven't been refactored and moved to `lib/`
and so are still using the old/fragile implementations written many
years ago.

I will be refactoring them more fully in the future, but for now the
`CACHE_DIR` issue has been resolved by more narrowly scoping where user
provided env vars are loaded.

Fixes #1886.
GUS-W-19443067.