Output a warning if an existing virtual environment is found
Since:
- Python virtual environments are specific to a single machine
  (they contain OS and architecture specific binaries, and hardcoded
  local paths), so should never be copied from machine to machine.
- Including them at best bloats the slug for no reason, and at worst
  can cause hard to debug errors if the package manager auto-detects
  them and uses them instead of the Python install managed by this
  buildpack (as seen in https://heroku.support/1604888)

A venv can end up in the app source either if someone forgets to
`.gitignore` it, but also if they use an alternative deployment means
that bypasses Git (such as Travis dpl), and haven't used another means
to exclude unwanted files.

This will be made an error fairly soon (once metrics confirm how
widespread the pattern is, or whether there are third-party buildpacks
that create a venv during the build).

GUS-W-19127248.