Improve env var and package manager tests (#1895)
The buildpack configures env vars for:
- pre/post-compile hooks
- Django collectstatic
- subsequent buildpacks (via the `export` file)
- app run-time (via `profile.d/` scripts)

The tests for all of these have been improved, by backporting some of
the new test matrix strategies from the Python CNB.

In particular, we now test the run-time and hooks env var values with
all four supported package managers (and not just when using pip).

In addition, we now test that Poetry and uv aren't installed at run-time.

GUS-W-19591873.