Deprecate implicit setup.py support (#1897)
Currently if an app has only a `setup.py` file and no Python package
manager files (such as `requirements.txt`, `Pipfile.lock`, `Poetry.lock`
or `uv.lock`), the buildpack will install the project using
`pip install --editable .`

However, there is nothing pip-specific about `setup.py` - the file is
not a package manager dependency file but instead a setuptools package
file, and could easily be installed by any package manager. In addition,
some projects may want to install this file in standard rather than
editable mode.

As such, it's best that the buildpack doesn't guess which package
manager and mode to use, and instead we require explicit configuration
from the user.

For example in the form of a `requirements.txt` file that references the
package definition, via this single line in the requirements file:

```
--editable .
```

That said, in general we recommend people don't use `setup.py` to declare
their dependencies, since:
(a) the file is deprecated in favour of `pyproject.toml`
(b) it's intended more for libraries rather than applications, and it's
    much less common (and practical) to list all transitive dependencies
    in a `setup.py`, meaning apps using the file typically have unpinned
    dependencies, which is a production reliability risk. (Apps should
    either be using a package manager that supports lockfiles, or else
    using one of the pip requirements files substitutes like pip-tools
    or `pip freeze` etc).

The existing metrics for this fallback show usage to be very low, so
we'll likely not wait long before converting this warning to an error
(particularly since fixing the error is a case of adding a single
line requirements file, so fairly simple).

GUS-W-19275438.